fix(optimum): allow non-fp32 dtypes for models that support them
Re-applies #800, reverted in #862 to unblock CI. Rebased onto current dev:
the platform.py helper imports moved into check_and_update_config, and the
converter tests moved under tests/optimum/.

The optimum path forced model_config.dtype to fp32 unconditionally. Only
optimum-rbln classes with _supports_non_fp32 = False actually need that, so
gate the override on forces_fp32_dtype() and let the rest keep the dtype the
caller asked for.

To keep vLLM and the compiled artefact in agreement:

- pass model_config.dtype to from_pretrained so the artefact is compiled with
  the requested dtype, and add dtype to the compile-cache path hash;
- read dtype back out of rbln_config.json and update model_config.dtype for a
  pre-compiled model, since the artefact is the source of truth;
- reject dtype inside additional_config's rbln_config, which optimum-rbln
  takes as a load argument rather than an rbln_config field.

Drop the now-redundant casts that assumed the fp32 override, and cast the
PaliGemma attention mask to the model's compute dtype instead of the compiled
torch_dtype.

+++ vllm_rbln/model_executor/models/optimum/qwen2_vl.py
@@ -355,9 +355,7 @@ class RBLNOptimumQwenVLForConditionalGeneration(
             caches = [c for c in cached_mm_outputs if spec.embeds_key in c]
             if not caches:
                 continue
-            mm[spec.embeds_key] = torch.cat(
-                [c[spec.embeds_key].to(self.dtype) for c in caches], dim=0
-            )
+            mm[spec.embeds_key] = torch.cat([c[spec.embeds_key] for c in caches], dim=0)
             mm[spec.grid_key] = torch.cat(
                 [c[spec.grid_key].to(torch.int64) for c in caches], dim=0
             )
